# fruit-store/src/test/resources/openapi.yaml
---
openapi: 3.0.3
info:
  title: fruit-store API
  version: 1.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value
paths:
  /:
    get:
      tags:
      - Web Resource
      responses:
        "200":
          description: OK
          content:
            text/html:
              schema:
                $ref: '#/components/schemas/TemplateInstance'
    post:
      tags:
      - Web Resource
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MultivaluedMapStringString'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateInstance'
  /api/fruits:
    get:
      tags:
      - Fruit Resource
      operationId: listFruits
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fruit'
    post:
      tags:
      - Fruit Resource
      operationId: addFruit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Fruit'
      responses:
        "200":
          description: OK
  /api/fruits/nutrition/{id}:
    get:
      tags:
      - Fruit Resource
      operationId: getNutrition
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/fruits/price/{id}:
    get:
      tags:
      - Fruit Resource
      operationId: getPriceUpdate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/fruits/{id}:
    get:
      tags:
      - Fruit Resource
      operationId: getFruitById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
    delete:
      tags:
      - Fruit Resource
      operationId: deleteFruit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /fruits:
    get:
      tags:
      - Web Resource
      responses:
        "200":
          description: OK
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/PublisherString'
components:
  schemas:
    Category:
      type: object
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
    Charset:
      type: object
      properties:
        name:
          type: string
        aliases:
          type: array
          items:
            type: string
        aliasSet:
          uniqueItems: true
          type: array
          items:
            type: string
        registered:
          type: boolean
    Expression:
      type: object
      properties:
        namespace:
          type: string
        parts:
          type: array
          items:
            $ref: '#/components/schemas/Part'
        literal:
          type: boolean
        literalValue:
          type: object
        origin:
          $ref: '#/components/schemas/Origin'
        generatedId:
          format: int32
          type: integer
    Fruit:
      type: object
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/Category'
        tags:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/Status'
        nutrition:
          $ref: '#/components/schemas/Nutrition'
        price:
          type: number
    Locale:
      type: object
      properties:
        language:
          type: string
        script:
          type: string
        country:
          type: string
        variant:
          type: string
        extensionKeys:
          uniqueItems: true
          type: array
          items:
            format: byte
            type: string
        unicodeLocaleAttributes:
          uniqueItems: true
          type: array
          items:
            type: string
        unicodeLocaleKeys:
          uniqueItems: true
          type: array
          items:
            type: string
        iSO3Language:
          type: string
        iSO3Country:
          type: string
        displayLanguage:
          type: string
        displayScript:
          type: string
        displayCountry:
          type: string
        displayVariant:
          type: string
        displayName:
          type: string
    MultivaluedMapStringString:
      type: object
      additionalProperties:
        type: array
        items:
          type: string
    Nutrition:
      type: object
      properties:
        id:
          format: int64
          type: integer
        calories:
          format: int32
          type: integer
        sugar:
          format: int32
          type: integer
    Origin:
      type: object
      properties:
        line:
          format: int32
          type: integer
        lineCharacterStart:
          format: int32
          type: integer
        lineCharacterEnd:
          format: int32
          type: integer
        templateId:
          type: string
        templateGeneratedId:
          type: string
        variant:
          type: object
          allOf:
          - $ref: '#/components/schemas/Variant'
          nullable: true
    ParameterDeclaration:
      type: object
      properties:
        typeInfo:
          type: string
        key:
          type: string
        defaultValue:
          $ref: '#/components/schemas/Expression'
        origin:
          $ref: '#/components/schemas/Origin'
    Part:
      type: object
      properties:
        name:
          type: string
        typeInfo:
          type: string
        virtualMethod:
          type: boolean
    PublisherString:
      type: object
    Status:
      enum:
      - SOLD
      - PENDING
      - AVAILABLE
      type: string
    Template:
      type: object
      properties:
        expressions:
          type: array
          items:
            $ref: '#/components/schemas/Expression'
        generatedId:
          type: string
        id:
          type: string
        variant:
          type: object
          allOf:
          - $ref: '#/components/schemas/Variant'
          nullable: true
        parameterDeclarations:
          type: array
          items:
            $ref: '#/components/schemas/ParameterDeclaration'
        fragment:
          type: boolean
    TemplateInstance:
      type: object
      properties:
        timeout:
          format: int64
          type: integer
        template:
          $ref: '#/components/schemas/Template'
    Variant:
      type: object
      properties:
        locale:
          $ref: '#/components/schemas/Locale'
        contentType:
          type: string
        encoding:
          $ref: '#/components/schemas/Charset'
        hashCode:
          format: int32
          type: integer
        mediaType:
          type: string
        charset:
          $ref: '#/components/schemas/Charset'